requirements.txt
PyMuPDF==1.27.1
python-dotenv
azure-ai-documentintelligence
rapidfuzz
fastapi
pydantic
httpx
uvicorn
motor
